Move API concerns from pretix/__init__.py to api/pretix/types.py
- Move Strawberry GraphQL types (CreateOrderInput, CreateOrderErrors, etc.) from pretix/__init__.py to api/pretix/types.py
- Update imports in api/orders/mutations.py and test files
- Add __all__ export list to pretix/__init__.py for backward compatibility
- Reduce pretix/__init__.py from 686 to 497 lines (189 lines removed)
- Keep core Pretix API client functions in pretix/__init__.py

This separation improves code organization by moving GraphQL API types
to the api/ directory where they belong, while keeping the Pretix HTTP
client logic in the pretix/ package.
